+++ app/components/CommentBox/index.js
@@ -1,0 +1,64 @@
+import React, { Component } from 'react'
+import CommentList from '../../components/CommentList'
+import CommentForm from '../../components/CommentForm'
+import $ from 'jquery'
+
+class CommentBox extends Component {
+  constructor(props) {
+    super(props)
+    this.state = {data: []}
+  }
+
+  loadCommentsFromServer() {
+    $.ajax({
+      url: this.props.url,
+      dataType: 'json',
+      cache: false,
+      success: function(data) {
+        this.props.actions.showComments(data)
+      }.bind(this),
+      error: function(xhr, status, err) {
+        console.error(this.props.url, status, err.toString())
+      }.bind(this)
+    })
+  }
+
+  handleCommentSubmit(comment) {
+    const comments = this.state.data
+    // Optimistically set an id on the new comment. It will be replaced by an
+    // id generated by the server. In a production application you would likely
+    // not use Date.now() for this and would have a more robust system in place.
+    comment.id = Date.now()
+    this.props.actions.addComment(comment)
+    $.ajax({
+      url: this.props.url,
+      dataType: 'json',
+      type: 'POST',
+      data: comment,
+      success: function(data) {
+        this.props.actions.showComments(data)
+      }.bind(this),
+      error: function(xhr, status, err) {
+        this.props.actions.showComments(comments)
+        console.error(this.props.url, status, err.toString())
+      }.bind(this)
+    })
+  }
+
+  componentDidMount() {
+    this.loadCommentsFromServer()
+    setInterval(this.loadCommentsFromServer.bind(this), this.props.pollInterval)
+  }
+
+  render() {
+    return (
+      <div className="commentBox">
+        <h1>Comments</h1>
+        <CommentList data={this.props.data} />
+        <CommentForm onCommentSubmit={this.handleCommentSubmit.bind(this)} />
+      </div>
+    )
+  }
+}
+
+export default CommentBox
