+++ app/containers/App/index.js
@@ -1,80 +1,40 @@
 import React, { Component } from 'react'
-import CommentList from '../../components/CommentList'
-import CommentForm from '../../components/CommentForm'
-import $ from 'jquery'
+import CommentBox from '../../components/CommentBox'
 import { connect } from 'react-redux'
 
-class CommentBox extends Component {
-  constructor(props) {
-    super(props)
-    this.state = {data: []}
-  }
-
-  loadCommentsFromServer() {
-    $.ajax({
-      url: this.props.url,
-      dataType: 'json',
-      cache: false,
-      success: function(data) {
-        this.props.showComments(data)
-      }.bind(this),
-      error: function(xhr, status, err) {
-        console.error(this.props.url, status, err.toString())
-      }.bind(this)
-    })
-  }
-
-  handleCommentSubmit(comment) {
-    const comments = this.state.data
-    // Optimistically set an id on the new comment. It will be replaced by an
-    // id generated by the server. In a production application you would likely
-    // not use Date.now() for this and would have a more robust system in place.
-    comment.id = Date.now()
-    this.props.addComment(comment)
-    $.ajax({
-      url: this.props.url,
-      dataType: 'json',
-      type: 'POST',
-      data: comment,
-      success: function(data) {
-        this.props.showComments(data)
-      }.bind(this),
-      error: function(xhr, status, err) {
-        this.props.showComments(comments)
-        console.error(this.props.url, status, err.toString())
-      }.bind(this)
-    })
-  }
-
-  componentDidMount() {
-    this.loadCommentsFromServer()
-    setInterval(this.loadCommentsFromServer.bind(this), this.props.pollInterval)
-  }
-
+class App extends Component {
   render() {
+    const { data, actions } = this.props
     return (
-      <div className="commentBox">
-        <h1>Comments</h1>
-        <CommentList data={this.props.data} />
-        <CommentForm onCommentSubmit={this.handleCommentSubmit.bind(this)} />
+      <div>
+        <CommentBox
+          data={data}
+          actions={actions}
+          url="http://localhost:3000/api/comments"
+          pollInterval={2000}
+        />
       </div>
     )
   }
 }
 
 const mapStateToProps = (state) => {
-  return {data: state}
+  return {
+    data: state
+  }
 }
 
 const mapDispatchToProps = (dispatch) => {
   return {
-    showComments(comments) {
-      dispatch({type: 'show_comments', comments: comments})
-    },
-    addComment(comment) {
-      dispatch({type: 'add_comment', comment: comment})
+    actions: {
+      showComments(comments) {
+        dispatch({type: 'show_comments', comments: comments})
+      },
+      addComment(comment) {
+        dispatch({type: 'add_comment', comment: comment})
+      }
     }
   }
 }
 
-export default connect(mapStateToProps, mapDispatchToProps)(CommentBox)
+export default connect(mapStateToProps, mapDispatchToProps)(App)
